unify fast and sync - change it regular sync assert
