+++ src/it/scala/io/iohk/ethereum/sync/util/RegularSyncItSpecUtils.scala
@@ -1,52 +1,28 @@
 package io.iohk.ethereum.sync.util
 
-import java.nio.file.Files
-import java.util.concurrent.atomic.AtomicReference
-
-import akka.actor.{ActorRef, ActorSystem}
-import akka.testkit.TestProbe
-import akka.util.{ByteString, Timeout}
+import akka.actor.ActorRef
+import akka.util.ByteString
 import cats.effect.Resource
 import io.iohk.ethereum.Mocks.MockValidatorsAlwaysSucceed
+import io.iohk.ethereum.blockchain.sync.PeersClient
 import io.iohk.ethereum.blockchain.sync.regular.BlockBroadcasterActor.BroadcastBlock
-import io.iohk.ethereum.blockchain.sync.regular.{BlockBroadcasterActor, RegularSync}
-import io.iohk.ethereum.blockchain.sync.{BlockBroadcast, BlockchainHostActor, PeersClient, TestSyncConfig}
+import io.iohk.ethereum.blockchain.sync.regular.RegularSync
 import io.iohk.ethereum.consensus.ethash.{EthashConfig, EthashConsensus}
 import io.iohk.ethereum.consensus.{ConsensusConfig, FullConsensusConfig, ethash}
-import io.iohk.ethereum.crypto.kec256
-import io.iohk.ethereum.db.components.{RocksDbDataSourceComponent, Storages}
-import io.iohk.ethereum.db.dataSource.{RocksDbConfig, RocksDbDataSource}
-import io.iohk.ethereum.db.storage.pruning.{ArchivePruning, PruningMode}
-import io.iohk.ethereum.db.storage.{AppStateStorage, Namespaces}
 import io.iohk.ethereum.domain._
 import io.iohk.ethereum.ledger._
-import io.iohk.ethereum.mpt.{HashNode, MerklePatriciaTrie, MptNode, MptTraversals}
-import io.iohk.ethereum.network.EtcPeerManagerActor.PeerInfo
-import io.iohk.ethereum.network.PeerManagerActor.{FastSyncHostConfiguration, PeerConfiguration}
-import io.iohk.ethereum.network.discovery.Node
-import io.iohk.ethereum.network.discovery.PeerDiscoveryManager.{DiscoveredNodesInfo, DiscoveryNodeInfo}
-import io.iohk.ethereum.network.handshaker.{EtcHandshaker, EtcHandshakerConfiguration, Handshaker}
-import io.iohk.ethereum.network.p2p.EthereumMessageDecoder
 import io.iohk.ethereum.network.p2p.messages.CommonMessages.NewBlock
-import io.iohk.ethereum.network.rlpx.AuthHandshaker
-import io.iohk.ethereum.network.rlpx.RLPxConnectionHandler.RLPxConfiguration
-import io.iohk.ethereum.network.{EtcPeerManagerActor, ForkResolver, KnownNodesManager, PeerEventBusActor, PeerManagerActor, ServerActor}
-import io.iohk.ethereum.nodebuilder.{PruningConfigBuilder, SecureRandomBuilder, ShutdownHookBuilder, VmSetup}
+import io.iohk.ethereum.nodebuilder.{ShutdownHookBuilder, VmSetup}
 import io.iohk.ethereum.ommers.OmmersPool
-import io.iohk.ethereum.sync.util.SyncCommonItSpec._
 import io.iohk.ethereum.sync.util.SyncCommonItSpecUtils.FakePeerCustomConfig.defaultConfig
 import io.iohk.ethereum.sync.util.SyncCommonItSpecUtils._
 import io.iohk.ethereum.transactions.PendingTransactionsManager
-import io.iohk.ethereum.utils.ServerStatus.Listening
 import io.iohk.ethereum.utils._
 import io.iohk.ethereum.vm.EvmConfig
-import io.iohk.ethereum.{Fixtures, Timeouts}
 import monix.eval.Task
 
-import scala.annotation.tailrec
 import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.{FiniteDuration, _}
-import scala.util.Try
+import scala.concurrent.duration._
 object RegularSyncItSpecUtils {
 
   object ShutdownHookBuilder extends ShutdownHookBuilder with Logger
@@ -58,181 +34,7 @@ object RegularSyncItSpecUtils {
   object ValidatorsExecutorAlwaysSucceed extends ValidatorsExecutorAlwaysSucceed
 
   class FakePeer(peerName: String, fakePeerCustomConfig: FakePeerCustomConfig)
-    extends SecureRandomBuilder
-      with TestSyncConfig {
-    implicit val akkaTimeout: Timeout = Timeout(5.second)
-
-    val config = Config.config
-
-    import scala.language.postfixOps
-
-    implicit val system = ActorSystem(peerName)
-
-    val peerDiscoveryManager = TestProbe().ref
-
-    val nodeKey = io.iohk.ethereum.crypto.generateKeyPair(secureRandom)
-
-    private val nodeStatus =
-      NodeStatus(
-        key = nodeKey,
-        serverStatus = ServerStatus.NotListening,
-        discoveryStatus = ServerStatus.NotListening
-      )
-
-    lazy val tempDir = Files.createTempDirectory("temp-regular-sync")
-
-    def getRockDbTestConfig(dbPath: String) = {
-      new RocksDbConfig {
-        override val createIfMissing: Boolean = true
-        override val paranoidChecks: Boolean = false
-        override val path: String = dbPath
-        override val maxThreads: Int = 1
-        override val maxOpenFiles: Int = 32
-        override val verifyChecksums: Boolean = false
-        override val levelCompaction: Boolean = true
-        override val blockSize: Long = 16384
-        override val blockCacheSize: Long = 33554432
-      }
-    }
-
-    sealed trait LocalPruningConfigBuilder extends PruningConfigBuilder {
-      override lazy val pruningMode: PruningMode = ArchivePruning
-    }
-
-    lazy val nodeStatusHolder = new AtomicReference(nodeStatus)
-    lazy val storagesInstance = new RocksDbDataSourceComponent
-      with LocalPruningConfigBuilder
-      with Storages.DefaultStorages {
-      override lazy val dataSource: RocksDbDataSource =
-        RocksDbDataSource(getRockDbTestConfig(tempDir.toAbsolutePath.toString), Namespaces.nsSeq)
-    }
-    lazy val blockchainConfig = Config.blockchains.blockchainConfig
-
-    /**
-      * Default persist interval is 20s, which is too long for tests. As in all tests we treat peer as connected when
-      * it is persisted in storage.
-      */
-    lazy val knownNodesManagerConfig =
-      KnownNodesManager.KnownNodesManagerConfig(config).copy(persistInterval = 1.seconds)
-
-    lazy val knownNodesManager = system.actorOf(
-      KnownNodesManager.props(
-        knownNodesManagerConfig,
-        storagesInstance.storages.knownNodesStorage
-      )
-    )
-
-    val bl = BlockchainImpl(storagesInstance.storages)
-
-    val genesis = Block(
-      Fixtures.Blocks.Genesis.header.copy(stateRoot = ByteString(MerklePatriciaTrie.EmptyRootHash)),
-      Fixtures.Blocks.Genesis.body
-    )
-
-    bl.save(genesis, Seq(), genesis.header.difficulty, saveAsBestBlock = true)
-
-    lazy val nh = nodeStatusHolder
-
-    val peerConf = new PeerConfiguration {
-      override val fastSyncHostConfiguration: FastSyncHostConfiguration = new FastSyncHostConfiguration {
-        val maxBlocksHeadersPerMessage: Int = fakePeerCustomConfig.hostConfig.maxBlocksHeadersPerMessage
-        val maxBlocksBodiesPerMessage: Int = fakePeerCustomConfig.hostConfig.maxBlocksBodiesPerMessage
-        val maxReceiptsPerMessage: Int = fakePeerCustomConfig.hostConfig.maxReceiptsPerMessage
-        val maxMptComponentsPerMessage: Int = fakePeerCustomConfig.hostConfig.maxMptComponentsPerMessage
-      }
-      override val rlpxConfiguration: RLPxConfiguration = new RLPxConfiguration {
-        override val waitForTcpAckTimeout: FiniteDuration = Timeouts.normalTimeout
-        override val waitForHandshakeTimeout: FiniteDuration = Timeouts.normalTimeout
-      }
-      override val waitForHelloTimeout: FiniteDuration = 3 seconds
-      override val waitForStatusTimeout: FiniteDuration = 30 seconds
-      override val waitForChainCheckTimeout: FiniteDuration = 15 seconds
-      override val connectMaxRetries: Int = 3
-      override val connectRetryDelay: FiniteDuration = 1 second
-      override val disconnectPoisonPillTimeout: FiniteDuration = 3 seconds
-      override val maxOutgoingPeers = 10
-      override val maxIncomingPeers = 5
-      override val maxPendingPeers = 5
-      override val networkId: Int = 1
-
-      override val updateNodesInitialDelay: FiniteDuration = 5.seconds
-      override val updateNodesInterval: FiniteDuration = 20.seconds
-      override val shortBlacklistDuration: FiniteDuration = 1.minute
-      override val longBlacklistDuration: FiniteDuration = 3.minutes
-    }
-
-    lazy val peerEventBus = system.actorOf(PeerEventBusActor.props, "peer-event-bus")
-
-    private val handshakerConfiguration: EtcHandshakerConfiguration =
-      new EtcHandshakerConfiguration {
-        override val forkResolverOpt: Option[ForkResolver] = None
-        override val nodeStatusHolder: AtomicReference[NodeStatus] = nh
-        override val peerConfiguration: PeerConfiguration = peerConf
-        override val blockchain: Blockchain = bl
-        override val appStateStorage: AppStateStorage = storagesInstance.storages.appStateStorage
-      }
-
-    lazy val handshaker: Handshaker[PeerInfo] = EtcHandshaker(handshakerConfiguration)
-
-    lazy val authHandshaker: AuthHandshaker = AuthHandshaker(nodeKey, secureRandom)
-
-    lazy val peerManager: ActorRef = system.actorOf(
-      PeerManagerActor.props(
-        peerDiscoveryManager,
-        Config.Network.peer,
-        peerEventBus,
-        knownNodesManager,
-        handshaker,
-        authHandshaker,
-        EthereumMessageDecoder
-      ),
-      "peer-manager"
-    )
-
-    lazy val etcPeerManager: ActorRef = system.actorOf(
-      EtcPeerManagerActor.props(peerManager, peerEventBus, storagesInstance.storages.appStateStorage, None),
-      "etc-peer-manager"
-    )
-
-    val blockchainHost: ActorRef =
-      system.actorOf(BlockchainHostActor.props(bl, peerConf, peerEventBus, etcPeerManager), "blockchain-host")
-
-    lazy val server: ActorRef = system.actorOf(ServerActor.props(nodeStatusHolder, peerManager), "server")
-
-    val listenAddress = randomAddress()
-
-    lazy val node =
-      DiscoveryNodeInfo(
-        Node(ByteString(nodeStatus.nodeId), listenAddress.getAddress, listenAddress.getPort, listenAddress.getPort),
-        1
-      )
-
-    lazy val vmConfig = VmConfig(Config.config)
-
-    lazy val validators = new MockValidatorsAlwaysSucceed
-
-    val testSyncConfig = syncConfig.copy(
-      minPeersToChoosePivotBlock = 1,
-      peersScanInterval = 5.milliseconds,
-      blockHeadersPerRequest = 200,
-      blockBodiesPerRequest = 50,
-      receiptsPerRequest = 50,
-      fastSyncThrottle = 10.milliseconds,
-      startRetryInterval = 50.milliseconds,
-      nodesPerRequest = 200,
-      maxTargetDifference = 1,
-      syncRetryInterval = 50.milliseconds
-    )
-
-    lazy val broadcaster = new BlockBroadcast(etcPeerManager, testSyncConfig)
-
-    lazy val broadcasterActor = system.actorOf(
-      BlockBroadcasterActor.props(broadcaster, peerEventBus, etcPeerManager, testSyncConfig, system.scheduler)
-    )
-
-    def startRegularSync(): Task[Unit] = Task {
-      regularSync ! RegularSync.Start
-    }
+    extends CommonFakePeer(peerName, fakePeerCustomConfig) {
 
     def buildEthashConsensus(): ethash.EthashConsensus = {
       val consensusConfig: ConsensusConfig = ConsensusConfig(Config.config)(ShutdownHookBuilder)
@@ -275,6 +77,10 @@ object RegularSyncItSpecUtils {
       )
     )
 
+    def startRegularSync(): Task[Unit] = Task {
+      regularSync ! RegularSync.Start
+    }
+
     def broadcastBlock(blockNumber: Option[Int] = None)(updateWorldForBlock: (BigInt, InMemoryWorldStateProxy) => InMemoryWorldStateProxy): Task[Unit] = {
       Task( blockNumber match {
         case Some(bNumber) => bl.getBlockByNumber(bNumber).getOrElse(throw new RuntimeException(s"block by number: $bNumber doesn't exist"))
@@ -301,6 +107,13 @@ object RegularSyncItSpecUtils {
       regularSync ! RegularSync.MinedBlock(newBlock)
     }
 
+    def mineNewBlocks(delay: FiniteDuration, nBlocks: Int)(updateWorldForBlock: (BigInt, InMemoryWorldStateProxy) => InMemoryWorldStateProxy): Task[Unit] = {
+        if(nBlocks > 0) {
+          mineNewBlock()(updateWorldForBlock).delayExecution(delay)
+            .flatMap(_ => mineNewBlocks(delay, nBlocks - 1)(updateWorldForBlock))
+        } else Task(())
+    }
+
     private def getMptForBlock(block: Block) = {
       bl.getWorldStateProxy(
         blockNumber = block.number,
@@ -315,46 +128,6 @@ object RegularSyncItSpecUtils {
       broadcasterActor ! BroadcastBlock(NewBlock(block, td))
     }
 
-    def getCurrentState(): BlockchainState = {
-      val bestBlock = bl.getBestBlock()
-      val currentWorldState = getMptForBlock(bestBlock)
-      val currentTd = bl.getTotalDifficultyByHash(bestBlock.hash).get
-      BlockchainState(bestBlock, currentWorldState, currentTd)
-    }
-
-    def startPeer(): Task[Unit] = {
-      for {
-        _ <- Task {
-          peerManager ! PeerManagerActor.StartConnecting
-          server ! ServerActor.StartServer(listenAddress)
-        }
-        _ <- retryUntilWithDelay(Task(nodeStatusHolder.get()), 1.second, 5) { status =>
-          status.serverStatus == Listening(listenAddress)
-        }
-      } yield ()
-    }
-
-    def shutdown(): Task[Unit] = {
-      for {
-        _ <- Task.deferFuture(system.terminate())
-        _ <- Task(storagesInstance.dataSource.destroy())
-      } yield ()
-    }
-
-    def connectToPeers(nodes: Set[DiscoveryNodeInfo]): Task[Unit] = {
-      for {
-        _ <- Task {
-          peerManager ! DiscoveredNodesInfo(nodes)
-        }
-        _ <- retryUntilWithDelay(Task(storagesInstance.storages.knownNodesStorage.getKnownNodes()), 1.second, 5) {
-          knownNodes =>
-            val requestedNodes = nodes.map(_.node.id)
-            val currentNodes = knownNodes.map(Node.fromUri).map(_.id)
-            requestedNodes.subsetOf(currentNodes)
-        }
-      } yield ()
-    }
-
     private def createChildBlock(parent: Block, parentTd: BigInt, parentWorld: InMemoryWorldStateProxy, plusDifficulty: BigInt = 0)(
       updateWorldForBlock: (BigInt, InMemoryWorldStateProxy) => InMemoryWorldStateProxy
     ): (Block, BigInt, InMemoryWorldStateProxy) = {
@@ -367,70 +140,6 @@ object RegularSyncItSpecUtils {
       )
       val newTd = newBlock.header.difficulty + parentTd
       (newBlock, newTd, parentWorld)
-    }
-
-    def importBlocksUntil(
-                           n: BigInt
-                         )(updateWorldForBlock: (BigInt, InMemoryWorldStateProxy) => InMemoryWorldStateProxy): Task[Unit] = {
-      Task(bl.getBestBlock()).flatMap { block =>
-        if (block.number >= n) {
-          Task(())
-        } else {
-          Task {
-            val currentTd = bl.getTotalDifficultyByHash(block.hash).get
-            val currentWolrd = getMptForBlock(block)
-            val (newBlock, newTd, newWorld) = createChildBlock(block, currentTd, currentWolrd)(updateWorldForBlock)
-            bl.save(newBlock, Seq(), newTd, saveAsBestBlock = true)
-            bl.persistCachedNodes()
-            broadcastBlock(newBlock, newTd)
-          }.flatMap(_ => importBlocksUntil(n)(updateWorldForBlock))
-        }
-      }
-    }
-
-    // Reads whole trie into memory, if the trie lacks nodes in storage it will be None
-    def getBestBlockTrie(): Option[MptNode] = {
-      Try {
-        val bestBlock = bl.getBestBlock()
-        val bestStateRoot = bestBlock.header.stateRoot
-        MptTraversals.parseTrieIntoMemory(
-          HashNode(bestStateRoot.toArray),
-          storagesInstance.storages.stateStorage.getBackingStorage(bestBlock.number)
-        )
-      }.toOption
-    }
-
-    def containsExpectedDataUpToAccountAtBlock(n: BigInt, blockNumber: BigInt): Boolean = {
-      @tailrec
-      def go(i: BigInt): Boolean = {
-        if (i >= n) {
-          true
-        } else {
-          val expectedBalance = i
-          val accountAddress = Address(i)
-          val accountExpectedCode = ByteString(i.toByteArray)
-          val codeHash = kec256(accountExpectedCode)
-          val accountExpectedStorageAddresses = (i until i + 20).toList
-          val account = bl.getAccount(accountAddress, blockNumber).get
-          val code = bl.getEvmCodeByHash(codeHash).get
-          val storedData = accountExpectedStorageAddresses.map { addr =>
-            ByteUtils.toBigInt(bl.getAccountStorageAt(account.storageRoot, addr, ethCompatibleStorage = true))
-          }
-          val haveAllStoredData = accountExpectedStorageAddresses.zip(storedData).forall { case (address, value) =>
-            address == value
-          }
-
-          val dataIsCorrect =
-            account.balance.toBigInt == expectedBalance && code == accountExpectedCode && haveAllStoredData
-          if (dataIsCorrect) {
-            go(i + 1)
-          } else {
-            false
-          }
-        }
-      }
-
-      go(0)
     }
   }
